massive_options_smoke/requirements.txt
requests
python-dotenv
pandas
boto3

